Refactor and consolidate vcfg management
Organized vcfg configuration handling by removing redundant cv.vcfg files. Introduced a more general Vcfg class and updated related code to accommodate new structure, enhancing maintainability and clarity.
