+++ experiment.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass, field
+from typing import Any
 
 
 @dataclass
@@ -11,16 +12,21 @@ class Block:
     definition: list[list[int]] = field(default_factory=list)
 
 @dataclass
+class Vcfg:
+    name: str
+    technique: str
+    configs: Any
+
+@dataclass
 class Gcode:
     name: str
     file: str
 
 class Experiment:
-    def __init__(self, solution, block, technique, vcfg, gcode):
+    def __init__(self, solution, block, vcfg, gcode):
         self.solution: str = solution
         self.block: Block = block
-        self.technique: str = technique
-        self.vcfg = vcfg
+        self.vcfg: Vcfg = vcfg
         self.gcode: Gcode = gcode
 
     def tile_block(self):
@@ -39,6 +45,6 @@ class Experiment:
                                       self.block.num_cols, new_start_row, new_start_col, self.block.definition)
 
     def __str__(self):
-        return (f'{self.solution[:24]:<25} Block: {self.block.name[:6]:<7} Mode: {self.technique[:6]:<7} '
+        return (f'{self.solution[:24]:<25} Block: {self.block.name[:6]:<7} Mode: {self.vcfg.technique[:4]:<5} '
                 f'Vcfg: {self.vcfg.name[:6]:<7} Well: {self.gcode.name[:6]:<7}')
