+++ fileio.py
@@ -1,12 +1,15 @@
 import ast
 import os
 import experiment
+import technique_fields
 
-def from_file(file_path, file_type):
+
+def from_file(file_path, file_type = None):
     with open(file_path, 'r') as file:
-        first_line = file.readline().strip()
-        if first_line != f"[{file_type} Config]":
-            raise ValueError(f"Invalid file format: first line must be [{file_type} Config]")
+        if file_type is not None:
+            first_line = file.readline().strip()
+            if first_line != f"[{file_type} Config]":
+                raise ValueError(f"Invalid file format: first line must be [{file_type} Config]")
         values = []
         name = os.path.basename(file_path).split('.')[0]
         values.append(name)
@@ -23,8 +26,9 @@ def from_folder(path, suffix):
             try:
                 if suffix == ".block":
                     file = experiment.Block(*from_file(file_path, 'Block'))
-                elif suffix == ".cv.vcfg":
-                    file = experiment.CV(*from_file(file_path, 'Cyclic Voltammetry'))
+                elif suffix == ".vcfg":
+                    values = from_file(file_path)
+                    file = experiment.Vcfg(values[0], values[1], technique_fields.CV(*values[2:]))
                 elif suffix == ".gcode":
                     filename = os.path.basename(file_path)
                     file = experiment.Gcode(filename.split('.')[0], filename)
