--- vcfgs/cv/oct test.cv.vcfg
@@ -1,10 +1,0 @@
-[Cyclic Voltammetry Config]
-v_start = 0.5
-v_end = 1.3
-v_2 = 0.2
-v_final = 1.0
-scan rate (V/s) = 0.05
-voltage step = 0.001
-cycles = 5
-begin_measuring_I = 0.5
-End_measuring_I = 1.0
