+++ vcfgs/test2.cv.vcfg
@@ -1,0 +1,10 @@
+technique = "CV"
+vs_init = [False, False, False, False, False]
+v_step = [0.6, 1.9, 0.0, 0.6, 0.6]
+scan_rate = [800.0, 800.0, 800.0, 800.0, 800.0]
+scan_number = 2
+record_de = 0.01
+average_de = True
+n_cycles = 25
+begin_i = 0.98
+end_i = 1.0
